user-svc/src/requirements/dev.txt
aniso8601==7.0.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
coverage==4.5.3
dnspython==1.16.0
eventlet==0.25.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
flask-mongoengine==0.9.5
flask-restplus==0.12.1
Flask-WTF==0.14.2
greenlet==0.4.15
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.18.0
mongomock==3.17.0
monotonic==1.5
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
pyrsistent==0.15.2
pytz==2019.1
sentinels==1.0.0
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
